[Vulkan][Refactor] Pull out vulkan initialization into VulkanInstance and VulkanDevice (#8188)
* [Vulkan][Refactor] Broke out VkInstance setup/teardown into managed class.

- Previously, the VkInstance was directly owned by the
  VulkanDeviceAPI.  Now, VulkanDeviceAPI owns a
  tvm::runtime::vulkan::VulkanInstance that does setup/teardown of the
  VkInstance.  This way, the teardown is done even if a later
  initialization step throws an exception.

* [Vulkan] Renamed VulkanContext to VulkanDevice

Renaming to match with the tvm.context to tvm.device rename.

* [Vulkan][Refactor] Extracted VulkanDevice initialization into VulkanDevice class

* [Vulkan][Refactor] Removed the VkPhysicalDeviceProperties member variable from VulkanDevice

- Now that there is a separate VulkanDeviceProperties class, the
  redundant VkPhysicalDeviceProperties can be removed.
